pyproject.toml
[tool.poetry]
name = "controller"
version = "0.1.0"
description = "Controller & UI"
authors = ["Taisei <st22182ti@gm.ibaraki-ct.ac.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.14"
uvicorn = "^0.34.3"
jinja2 = "^3.1.6"
google-genai = "^1.23.0"
dotenv = "^0.9.9"
toml = "^0.10.2"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
